Re-organize chart service
